screencast: Simplify shutdown/error flow
Shutdown can be initiated from multiple directions - pipewire, dbus and
the display server. To synchronize everything, various flags were used
to defer cleanup until all parties had time to react. This deferred
shutdown requires checking these flags in all relevant locations and
increases the risk of an accidental UAF.

Replace the flags with explicit and immediate shutdown, ensuring that no
further events will be received after the shutdown is initiated.
