+++ src/screencast/screencast.c
@@ -73,7 +73,6 @@ void xdpw_screencast_instance_init(struct xdpw_screencast_context *ctx,
 	cast->refcount = 1;
 	cast->node_id = SPA_ID_INVALID;
 	cast->avoid_dmabufs = false;
-	cast->teardown = false;
 	wl_list_init(&cast->buffer_list);
 	logprint(INFO, "xdpw: screencast instance %p has %d references", cast, cast->refcount);
 	wl_list_insert(&ctx->screencast_instances, &cast->link);
@@ -82,6 +81,25 @@ void xdpw_screencast_instance_init(struct xdpw_screencast_context *ctx,
 }
 
 void xdpw_screencast_instance_destroy(struct xdpw_screencast_instance *cast) {
+	struct xdpw_timer *timer, *ttmp;
+	wl_list_for_each_safe(timer, ttmp, &cast->ctx->state->timers, link) {
+		if (timer->user_data == cast) {
+			xdpw_destroy_timer(timer);
+		}
+	}
+	struct xdpw_session *sess, *stmp;
+	wl_list_for_each_safe(sess, stmp, &cast->ctx->state->xdpw_sessions, link) {
+		if (sess->screencast_data.screencast_instance == cast) {
+			wl_list_remove(&sess->link);
+			wl_list_init(&sess->link);
+			sess->screencast_data.screencast_instance = NULL;
+			cast->refcount--;
+			xdpw_session_destroy(sess);
+		}
+	}
+
+	xdpw_wlr_session_close(cast);
+
 	assert(cast->refcount == 0); // Fails assert if called by screencast_finish
 	logprint(DEBUG, "xdpw: destroying cast instance");
 
@@ -102,15 +120,6 @@ void xdpw_screencast_instance_destroy(struct xdpw_screencast_instance *cast) {
 	xdpw_buffer_constraints_finish(&cast->current_constraints);
 	xdpw_buffer_constraints_finish(&cast->pending_constraints);
 	free(cast);
-}
-
-void xdpw_screencast_instance_teardown(struct xdpw_screencast_instance *cast) {
-	struct xdpw_session *sess, *tmp;
-	wl_list_for_each_safe(sess, tmp, &cast->ctx->state->xdpw_sessions, link) {
-		if (sess->screencast_data.screencast_instance == cast) {
-			xdpw_session_destroy(sess);
-		}
-	}
 }
 
 bool setup_target(struct xdpw_screencast_context *ctx, struct xdpw_session *sess, struct xdpw_screencast_restore_data *data) {
